--- src/scripts/downstream-browsers.ts
@@ -1,1559 +1,0 @@
-type DownstreamBrowsers = {
-  browsers: {
-    [name: string]: {
-      name: string;
-      type: string;
-      upstream: string;
-      releases: {
-        [version: string]: {
-          engine: string;
-          engine_version: string;
-          status: string;
-          release_date: string;
-        };
-      };
-    };
-  };
-  lastUpdated: string;
-};
-
-const downstreamBrowsers: DownstreamBrowsers = {
-  browsers: {
-    ya_android: {
-      name: "Yandex Browser Mobile",
-      type: "mobile",
-      upstream: "chrome_android",
-      releases: {
-        "1.0": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "1.5": {
-          engine: "Blink",
-          engine_version: "22",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "1.6": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "1.7": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "1.20": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "2.5": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "3.2": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "4.6": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "5.3": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "5.4": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.4": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "9.6": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.5": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.4": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.5": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.7": {
-          engine: "Blink",
-          engine_version: "25",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "13.9": {
-          engine: "Blink",
-          engine_version: "28",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "13.10": {
-          engine: "Blink",
-          engine_version: "28",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "13.11": {
-          engine: "Blink",
-          engine_version: "28",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "13.12": {
-          engine: "Blink",
-          engine_version: "30",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "14.2": {
-          engine: "Blink",
-          engine_version: "32",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "14.4": {
-          engine: "Blink",
-          engine_version: "33",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "14.5": {
-          engine: "Blink",
-          engine_version: "34",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "14.7": {
-          engine: "Blink",
-          engine_version: "35",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "14.8": {
-          engine: "Blink",
-          engine_version: "36",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "14.10": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "14.12": {
-          engine: "Blink",
-          engine_version: "38",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "15.2": {
-          engine: "Blink",
-          engine_version: "40",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "15.4": {
-          engine: "Blink",
-          engine_version: "41",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "15.6": {
-          engine: "Blink",
-          engine_version: "42",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "15.7": {
-          engine: "Blink",
-          engine_version: "43",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "15.9": {
-          engine: "Blink",
-          engine_version: "44",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "15.10": {
-          engine: "Blink",
-          engine_version: "45",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "15.12": {
-          engine: "Blink",
-          engine_version: "46",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "16.2": {
-          engine: "Blink",
-          engine_version: "47",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "16.3": {
-          engine: "Blink",
-          engine_version: "47",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "16.4": {
-          engine: "Blink",
-          engine_version: "49",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "16.6": {
-          engine: "Blink",
-          engine_version: "50",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "16.7": {
-          engine: "Blink",
-          engine_version: "51",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "16.9": {
-          engine: "Blink",
-          engine_version: "52",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "16.10": {
-          engine: "Blink",
-          engine_version: "53",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "16.11": {
-          engine: "Blink",
-          engine_version: "54",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "17.1": {
-          engine: "Blink",
-          engine_version: "55",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "17.3": {
-          engine: "Blink",
-          engine_version: "56",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "17.4": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "17.6": {
-          engine: "Blink",
-          engine_version: "58",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "17.7": {
-          engine: "Blink",
-          engine_version: "59",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "17.9": {
-          engine: "Blink",
-          engine_version: "60",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "17.10": {
-          engine: "Blink",
-          engine_version: "61",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "17.11": {
-          engine: "Blink",
-          engine_version: "62",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "18.1": {
-          engine: "Blink",
-          engine_version: "63",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "18.2": {
-          engine: "Blink",
-          engine_version: "63",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "18.3": {
-          engine: "Blink",
-          engine_version: "64",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "18.4": {
-          engine: "Blink",
-          engine_version: "65",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "18.6": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "18.7": {
-          engine: "Blink",
-          engine_version: "67",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "18.9": {
-          engine: "Blink",
-          engine_version: "68",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "18.10": {
-          engine: "Blink",
-          engine_version: "69",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "18.11": {
-          engine: "Blink",
-          engine_version: "70",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.1": {
-          engine: "Blink",
-          engine_version: "71",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.3": {
-          engine: "Blink",
-          engine_version: "72",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.4": {
-          engine: "Blink",
-          engine_version: "73",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.5": {
-          engine: "Blink",
-          engine_version: "75",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.6": {
-          engine: "Blink",
-          engine_version: "75",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.7": {
-          engine: "Blink",
-          engine_version: "75",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.9": {
-          engine: "Blink",
-          engine_version: "76",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.10": {
-          engine: "Blink",
-          engine_version: "77",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.11": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "19.12": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "20.2": {
-          engine: "Blink",
-          engine_version: "79",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "20.3": {
-          engine: "Blink",
-          engine_version: "80",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "20.4": {
-          engine: "Blink",
-          engine_version: "81",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "20.6": {
-          engine: "Blink",
-          engine_version: "81",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "20.7": {
-          engine: "Blink",
-          engine_version: "83",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "20.8": {
-          engine: "Blink",
-          engine_version: "84",
-          status: "unknown",
-          release_date: "2020-09-02",
-        },
-        "20.9": {
-          engine: "Blink",
-          engine_version: "85",
-          status: "unknown",
-          release_date: "2020-09-27",
-        },
-        "20.11": {
-          engine: "Blink",
-          engine_version: "86",
-          status: "unknown",
-          release_date: "2020-11-11",
-        },
-        "20.12": {
-          engine: "Blink",
-          engine_version: "87",
-          status: "unknown",
-          release_date: "2020-12-20",
-        },
-        "21.1": {
-          engine: "Blink",
-          engine_version: "88",
-          status: "unknown",
-          release_date: "2021-12-31",
-        },
-        "21.2": {
-          engine: "Blink",
-          engine_version: "88",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "21.3": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2021-04-04",
-        },
-        "21.5": {
-          engine: "Blink",
-          engine_version: "90",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "21.6": {
-          engine: "Blink",
-          engine_version: "91",
-          status: "unknown",
-          release_date: "2021-09-28",
-        },
-        "21.8": {
-          engine: "Blink",
-          engine_version: "92",
-          status: "unknown",
-          release_date: "2021-09-28",
-        },
-        "21.9": {
-          engine: "Blink",
-          engine_version: "93",
-          status: "unknown",
-          release_date: "2021-09-29",
-        },
-        "21.11": {
-          engine: "Blink",
-          engine_version: "94",
-          status: "unknown",
-          release_date: "2021-10-29",
-        },
-        "22.1": {
-          engine: "Blink",
-          engine_version: "96",
-          status: "unknown",
-          release_date: "2021-12-31",
-        },
-        "22.3": {
-          engine: "Blink",
-          engine_version: "98",
-          status: "unknown",
-          release_date: "2022-03-25",
-        },
-        "22.4": {
-          engine: "Blink",
-          engine_version: "92",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "22.5": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2022-05-20",
-        },
-        "22.7": {
-          engine: "Blink",
-          engine_version: "102",
-          status: "unknown",
-          release_date: "2022-07-07",
-        },
-        "22.8": {
-          engine: "Blink",
-          engine_version: "104",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "22.9": {
-          engine: "Blink",
-          engine_version: "104",
-          status: "unknown",
-          release_date: "2022-08-27",
-        },
-        "22.11": {
-          engine: "Blink",
-          engine_version: "106",
-          status: "unknown",
-          release_date: "2022-11-11",
-        },
-        "23.1": {
-          engine: "Blink",
-          engine_version: "108",
-          status: "unknown",
-          release_date: "2023-01-10",
-        },
-        "23.3": {
-          engine: "Blink",
-          engine_version: "110",
-          status: "unknown",
-          release_date: "2023-03-26",
-        },
-        "23.5": {
-          engine: "Blink",
-          engine_version: "112",
-          status: "unknown",
-          release_date: "2023-05-19",
-        },
-        "23.7": {
-          engine: "Blink",
-          engine_version: "114",
-          status: "unknown",
-          release_date: "2023-07-06",
-        },
-        "23.9": {
-          engine: "Blink",
-          engine_version: "116",
-          status: "unknown",
-          release_date: "2023-09-13",
-        },
-        "23.11": {
-          engine: "Blink",
-          engine_version: "118",
-          status: "unknown",
-          release_date: "2023-11-15",
-        },
-        "24.1": {
-          engine: "Blink",
-          engine_version: "120",
-          status: "unknown",
-          release_date: "2024-01-18",
-        },
-        "24.2": {
-          engine: "Blink",
-          engine_version: "120",
-          status: "unknown",
-          release_date: "2024-03-25",
-        },
-        "24.4": {
-          engine: "Blink",
-          engine_version: "122",
-          status: "unknown",
-          release_date: "2024-03-27",
-        },
-        "24.6": {
-          engine: "Blink",
-          engine_version: "124",
-          status: "unknown",
-          release_date: "2024-06-04",
-        },
-        "24.7": {
-          engine: "Blink",
-          engine_version: "126",
-          status: "unknown",
-          release_date: "2024-07-18",
-        },
-        "24.9": {
-          engine: "Blink",
-          engine_version: "126",
-          status: "unknown",
-          release_date: "2024-10-01",
-        },
-        "24.10": {
-          engine: "Blink",
-          engine_version: "128",
-          status: "unknown",
-          release_date: "2024-10-11",
-        },
-        "24.12": {
-          engine: "Blink",
-          engine_version: "130",
-          status: "unknown",
-          release_date: "2024-11-30",
-        },
-        "25.2": {
-          engine: "Blink",
-          engine_version: "132",
-          status: "unknown",
-          release_date: "2025-04-24",
-        },
-        "25.3": {
-          engine: "Blink",
-          engine_version: "132",
-          status: "unknown",
-          release_date: "2025-04-23",
-        },
-        "25.4": {
-          engine: "Blink",
-          engine_version: "134",
-          status: "unknown",
-          release_date: "2025-04-23",
-        },
-        "99.9": {
-          engine: "Blink",
-          engine_version: "800",
-          status: "unknown",
-          release_date: "2025-05-27",
-        },
-      },
-    },
-    uc_android: {
-      name: "UC Browser Mobile",
-      type: "mobile",
-      upstream: "chrome_android",
-      releases: {
-        "10.5": {
-          engine: "Blink",
-          engine_version: "31",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.7": {
-          engine: "Blink",
-          engine_version: "31",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.8": {
-          engine: "Blink",
-          engine_version: "31",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.10": {
-          engine: "Blink",
-          engine_version: "31",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.0": {
-          engine: "Blink",
-          engine_version: "31",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.1": {
-          engine: "Blink",
-          engine_version: "40",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.2": {
-          engine: "Blink",
-          engine_version: "40",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.3": {
-          engine: "Blink",
-          engine_version: "40",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.4": {
-          engine: "Blink",
-          engine_version: "40",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.5": {
-          engine: "Blink",
-          engine_version: "40",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.6": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.8": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.9": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.0": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.1": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.2": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.3": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.4": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.5": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.6": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.7": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.8": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.9": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.10": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.11": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.12": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.13": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.14": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "13.0": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "13.1": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "13.2": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "13.3": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2020-09-09",
-        },
-        "13.4": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2021-09-28",
-        },
-        "13.5": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2023-08-25",
-        },
-        "13.6": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2023-12-17",
-        },
-        "13.7": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2023-06-24",
-        },
-        "13.8": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2022-04-30",
-        },
-        "13.9": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2022-05-18",
-        },
-        "15.0": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2022-08-24",
-        },
-        "15.1": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2022-11-11",
-        },
-        "15.2": {
-          engine: "Blink",
-          engine_version: "78",
-          status: "unknown",
-          release_date: "2023-04-23",
-        },
-        "15.3": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2023-03-17",
-        },
-        "15.4": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2023-10-25",
-        },
-        "15.5": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2023-08-22",
-        },
-        "16.0": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2023-08-24",
-        },
-        "16.1": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2023-10-15",
-        },
-        "16.2": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2023-12-09",
-        },
-        "16.3": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2024-03-08",
-        },
-        "16.4": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2024-10-03",
-        },
-        "16.5": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2024-05-30",
-        },
-        "16.6": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2024-07-23",
-        },
-        "17.0": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2024-08-24",
-        },
-        "17.1": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2024-09-26",
-        },
-        "17.2": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2024-11-29",
-        },
-        "17.3": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2025-01-07",
-        },
-        "17.4": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2025-02-26",
-        },
-        "17.5": {
-          engine: "Blink",
-          engine_version: "100",
-          status: "unknown",
-          release_date: "2025-04-08",
-        },
-        "17.6": {
-          engine: "Blink",
-          engine_version: "123",
-          status: "unknown",
-          release_date: "2025-05-15",
-        },
-      },
-    },
-    qq_android: {
-      name: "QQ Browser Mobile",
-      type: "mobile",
-      upstream: "chrome_android",
-      releases: {
-        "6.0": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "6.1": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "6.2": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "6.3": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "6.4": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "6.6": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "6.7": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "6.8": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "6.9": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.0": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.1": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.2": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.3": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.4": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.5": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.6": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.7": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.8": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "7.9": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.0": {
-          engine: "Blink",
-          engine_version: "37",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.1": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.2": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.3": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.4": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.5": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.6": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.7": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.8": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "8.9": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "9.1": {
-          engine: "Blink",
-          engine_version: "57",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "9.6": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "9.7": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "9.8": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.0": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.1": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.2": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.3": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.4": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.5": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "10.7": {
-          engine: "Blink",
-          engine_version: "66",
-          status: "unknown",
-          release_date: "2020-09-09",
-        },
-        "10.9": {
-          engine: "Blink",
-          engine_version: "77",
-          status: "unknown",
-          release_date: "2020-11-22",
-        },
-        "11.0": {
-          engine: "Blink",
-          engine_version: "77",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "11.2": {
-          engine: "Blink",
-          engine_version: "77",
-          status: "unknown",
-          release_date: "2021-01-30",
-        },
-        "11.3": {
-          engine: "Blink",
-          engine_version: "77",
-          status: "unknown",
-          release_date: "2021-03-31",
-        },
-        "11.7": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2021-11-02",
-        },
-        "11.9": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "unknown",
-        },
-        "12.0": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2021-11-04",
-        },
-        "12.1": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2021-11-05",
-        },
-        "12.2": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2021-12-07",
-        },
-        "12.5": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2022-04-07",
-        },
-        "12.7": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2022-05-21",
-        },
-        "12.8": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2022-06-30",
-        },
-        "12.9": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2022-07-26",
-        },
-        "13.0": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2022-08-15",
-        },
-        "13.1": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2022-09-10",
-        },
-        "13.2": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2022-10-26",
-        },
-        "13.3": {
-          engine: "Blink",
-          engine_version: "89",
-          status: "unknown",
-          release_date: "2022-11-09",
-        },
-        "13.4": {
-          engine: "Blink",
-          engine_version: "98",
-          status: "unknown",
-          release_date: "2023-04-26",
-        },
-        "13.5": {
-          engine: "Blink",
-          engine_version: "98",
-          status: "unknown",
-          release_date: "2023-02-06",
-        },
-        "13.6": {
-          engine: "Blink",
-          engine_version: "98",
-          status: "unknown",
-          release_date: "2023-02-09",
-        },
-        "13.7": {
-          engine: "Blink",
-          engine_version: "98",
-          status: "unknown",
-          release_date: "2023-04-21",
-        },
-        "13.8": {
-          engine: "Blink",
-          engine_version: "98",
-          status: "unknown",
-          release_date: "2023-04-21",
-        },
-        "14.0": {
-          engine: "Blink",
-          engine_version: "98",
-          status: "unknown",
-          release_date: "2023-12-12",
-        },
-        "14.1": {
-          engine: "Blink",
-          engine_version: "98",
-          status: "unknown",
-          release_date: "2023-07-16",
-        },
-        "14.2": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2023-10-14",
-        },
-        "14.3": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2023-09-13",
-        },
-        "14.4": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2023-10-31",
-        },
-        "14.5": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2023-11-12",
-        },
-        "14.6": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2023-12-24",
-        },
-        "14.7": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-01-18",
-        },
-        "14.8": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-03-04",
-        },
-        "14.9": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-04-09",
-        },
-        "15.0": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-04-17",
-        },
-        "15.1": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-05-18",
-        },
-        "15.2": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-10-24",
-        },
-        "15.3": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-07-28",
-        },
-        "15.4": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-09-07",
-        },
-        "15.5": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-09-24",
-        },
-        "15.6": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-10-24",
-        },
-        "15.7": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-12-03",
-        },
-        "15.8": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2024-12-11",
-        },
-        "15.9": {
-          engine: "Blink",
-          engine_version: "109",
-          status: "unknown",
-          release_date: "2025-02-01",
-        },
-      },
-    },
-  },
-  lastUpdated: "2025-05-28T14:13:03.911Z",
-};
-
-export default downstreamBrowsers;
