+++ src/scripts/import-local-legacy.ts
@@ -1,6 +1,7 @@
 //@ts-ignore
 import bcdBrowsers from "@mdn/browser-compat-data" assert { type: "json" };
-import otherBrowsers from "./downstream-browsers.js";
+//@ts-ignore
+import otherBrowsers from "../data/downstream-browsers.json" assert { type: "json" };
 import { features } from "web-features";
 
 export { bcdBrowsers, otherBrowsers, features };
